// src/types.ts
export type Action = {
    type: string;
    payload: any;
    meta?: {
        skipHistory?: boolean;
        [key: string]: any;
    }
};
export type ToDispatch = Action[];

export interface Operation {
    add: (...actions: Action[]) => void;
    clear: () => void;
    addDiff: (fn: DiffFactoryFn) => void;
    performDiff: (fn: DiffFactoryFn) => void;
    submit: () => void;
    state: ActionState;
}

export type CardinalDirection = "n" | "w" | "s" | "e";
export type IntercardinalDirection = "ne" | "nw" | "se" | "sw";

export type HSLColor = [number, number, number];
export type RGBColor = [number, number, number];
export type RGBAColor = [number, number, number, number];
// ...
export enum PropertyGroupName {
    Transform = 5000,
    Dimensions = 5001,
    Content = 5002,
    Structure = 5003,
    Modifiers = 5004,
    Shape = 5006,
    Fill = 5007,
    Stroke = 5008,

    // Modifiers
    ArrayModifier = 5005,
}

export enum CompoundPropertyName {
    Anchor = 1000,
    Scale = 1001,
    Position = 1002,
    ArrayModifier_Origin = 1003,
}

export enum PropertyName {
    // Transform Properties
    AnchorX = 0,
    AnchorY = 1,
    Scale = 2,
    ScaleX = 24,
    ScaleY = 25,
    PositionX = 3,
    PositionY = 4,
    Rotation = 5,
    Opacity = 6,

    // Rect properties
    Width = 7,
    Height = 8,

    // Look properties
    Fill = 9,
    StrokeColor = 10,
    StrokeWidth = 11,
    BorderRadius = 12,
    RGBAColor = 18,
    RGBColor = 23,

    // Ellipse properties
    OuterRadius = 13,
    InnerRadius = 14,

    // Array Modifier
    ArrayModifier_Count = 15,
    ArrayModifier_TransformBehavior = 16,
    ArrayModifier_RotationCorrection = 26,
    ArrayModifier_OriginX = 27,
    ArrayModifier_OriginY = 28,
    ArrayModifier_OriginBehavior = 29,

    // Shape Layer
    ShapeLayer_Path = 17,
    FillRule = 19,
    LineCap = 20,
    LineJoin = 21,
    MiterLimit = 22,
}
// ...
export type FillRule = "evenodd" | "nonzero";
export type LineCap = "butt" | "round" | "square";
export type LineJoin = "miter" | "round" | "bevel";